Restructure code (just moved functions and files, no changes)
