+++ app/src/main/java/com/kay/simpletaxapp/MainActivity.kt
@@ -4,11 +4,13 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.kay.simpletaxapp.ui.screen.MainContent
+import com.kay.simpletaxapp.ui.screen.MainScreen
 import com.kay.simpletaxapp.ui.theme.SimpleTaxAppTheme
 import com.kay.simpletaxapp.ui.viewmodel.TaxViewModel
 
@@ -21,26 +23,10 @@ class MainActivity : ComponentActivity() {
         setContent {
             SimpleTaxAppTheme {
                 // A surface container using the 'background' color from the theme
-                MyApp(taxViewModel = taxViewModel) {
-                    MainContent(taxViewModel = taxViewModel)
+                Surface(modifier = Modifier.fillMaxSize(), color = MaterialTheme.colors.background) {
+                    MainScreen(taxViewModel = taxViewModel)
                 }
             }
-        }
-    }
-}
-
-@Composable
-fun MyApp(
-    taxViewModel: TaxViewModel,
-    content: @Composable () -> Unit,
-) {
-    SimpleTaxAppTheme {
-        // A surface container using the 'background' color from the theme
-        Surface(
-            color = MaterialTheme.colors.background
-        ) {
-            // content()
-            MainContent(taxViewModel = taxViewModel)
         }
     }
 }
@@ -49,7 +35,7 @@ fun MyApp(
 @Composable
 fun DefaultPreview() {
     val taxViewModel = TaxViewModel()
-    MyApp(taxViewModel = taxViewModel) {
-        MainContent(taxViewModel = taxViewModel)
+    Surface(color = MaterialTheme.colors.background) {
+        MainScreen(taxViewModel = taxViewModel)
     }
 }
