+++ app/src/main/java/com/kay/simpletaxapp/ui/screen/MainScreen.kt
@@ -4,13 +4,13 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.kay.simpletaxapp.component.TaxForm
+import com.kay.simpletaxapp.ui.component.TaxForm
 import com.kay.simpletaxapp.ui.theme.LARGE_PADDING
 import com.kay.simpletaxapp.ui.viewmodel.TaxViewModel
 import com.kay.simpletaxapp.util.sliderToPercentage
 
 @Composable
-fun MainContent(taxViewModel: TaxViewModel) {
+fun MainScreen(taxViewModel: TaxViewModel) {
 
     val viewState = taxViewModel.viewState
     val sliderPercentage = sliderToPercentage(viewState.sliderPositionState)
@@ -19,11 +19,7 @@ fun MainContent(taxViewModel: TaxViewModel) {
         TaxForm(
             taxViewModel = taxViewModel,
             viewState = viewState,
-            sliderPositionState = viewState.sliderPositionState,
-            percentage = sliderPercentage
-            // totalSalaryAmountState = viewState.totalSalaryAmountState,
-            // totalIncomeAfterTax = viewState.totalIncomeAfterTax,
-            // taxPay = viewState.taxAmountState,
-        ) {}
+            percentage = sliderPercentage,
+        )
     }
 }
